Move customizable code under keymap dir
