+++ keyboards/cocot46plus/cocot46plus.h
@@ -38,41 +38,3 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
     { B30,   B31,   B32,   B33,   B34,   B35 },  \
     { KC_NO, KC_NO, B42, KC_NO, KC_NO,   B45 }  \
   }
-
-
-typedef union {
-    uint32_t raw;
-    struct {
-        uint8_t cpi_idx;
-        uint8_t scrl_div;
-        uint8_t rotation_angle;
-        int8_t scrl_inv;
-        bool scrl_mode;
-        report_mouse_t last_mouse;
-    };
-} cocot_config_t;
-
-
-
-extern cocot_config_t cocot_config;
-
-enum cocot_keycodes {
-
-    COCOT_SAFE_RANGE = SAFE_RANGE,
-    CPI_SW,
-    SCRL_SW,
-    ROT_R15,
-    ROT_L15,
-    SCRL_MO,
-    SCRL_TO,
-    SCRL_IN,
-
-};
-
-bool encoder_update_user(uint8_t index, bool clockwise);
-bool encoder_update_kb(uint8_t index, bool clockwise);
-bool cocot_get_scroll_mode(void);
-void cocot_set_scroll_mode(bool mode);
-
-void render_logo(void);
-void oled_write_layer_state(void);
