+++ keyboards/cocot46plus/config.h
@@ -44,40 +44,3 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 /* Limit the maximum layer number to 8 */
 #define LAYER_STATE_8BIT
-
-/* Rotary encoder */
-#define ENCODERS_PAD_A { D3 }
-#define ENCODERS_PAD_B { D2 }
-#define ENCODER_RESOLUTION 4
-
-#define TAPPING_FORCE_HOLD
-#define TAPPING_TERM 200
-
-/* ws2812 RGB LED */
-#define RGB_DI_PIN B6
-
-#ifdef RGBLIGHT_ENABLE
-    #define RGBLIGHT_EFFECT_BREATHING
-    #define RGBLIGHT_EFFECT_RAINBOW_MOOD
-    #define RGBLIGHT_EFFECT_RAINBOW_SWIRL
-    #define RGBLIGHT_EFFECT_SNAKE
-    #define RGBLIGHT_EFFECT_KNIGHT
-    #define RGBLIGHT_EFFECT_CHRISTMAS
-    #define RGBLIGHT_EFFECT_STATIC_GRADIENT
-    #define RGBLIGHT_EFFECT_RGB_TEST
-    #define RGBLIGHT_EFFECT_ALTERNATING
-    // #define RGBLIGHT_EFFECT_TWINKLE
-
-    #define RGBLED_NUM 12
-    #define RGBLIGHT_LIMIT_VAL 120
-    #define RGBLIGHT_HUE_STEP 10
-    #define RGBLIGHT_SAT_STEP 17
-    #define RGBLIGHT_VAL_STEP 17
-#endif
-
-#define ADNS5050_SCLK_PIN         B2
-#define ADNS5050_SDIO_PIN         B4
-#define ADNS5050_CS_PIN           B5
-
-#define POINTING_DEVICE_ROTATION_180
-#define OLED_FONT_H "keyboards/cocot46plus/glcdfont.c"
