+++ keyboards/cocot46plus/keymaps/default/rules.mk
@@ -1,1 +1,15 @@
 VIA_ENABLE = yes
+
+MOUSEKEY_ENABLE = yes
+
+EXTRAKEY_ENABLE = yes
+
+ENCODER_ENABLE = yes
+
+RGBLIGHT_ENABLE = yes
+
+POINTING_DEVICE_ENABLE = yes
+POINTING_DEVICE_DRIVER = adns5050
+
+OLED_ENABLE     = yes
+OLED_DRIVER     = SSD1306
